+++ src/UI/Texture2D.h
@@ -47,78 +47,15 @@ public:
 
 
 
-	Texture2D(int width, int height)
-		: width(width), height(height)
-	{
-		glGenTextures(1, &id);
-		GL::CheckError();
-
-		Bind();
-
-		glTexImage2D(GL_TEXTURE_2D, 0,
-			GL_RGBA, width, height, 0,
-			GL_RGBA, GL_UNSIGNED_BYTE, 0);
-		GL::CheckError();
-
-		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-		GL::CheckError();
-
-		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-		GL::CheckError();
-
-		Unbind();
-	}
-
-	~Texture2D()
-	{
-		glDeleteTextures(1, &id);
-		GL::CheckError();
-	}
+	Texture2D(int width, int height);
+	~Texture2D();
 
 
-	void Bind()
-	{
-		glBindTexture(GL_TEXTURE_2D, id);
-		GL::CheckError();
-	}
 
-	void Unbind()
-	{
-		glBindTexture(GL_TEXTURE_2D, 0);
-		GL::CheckError();
-	}
-
-	void WriteData(void* data)
-	{
-		Bind();
-
-		glTexSubImage2D(GL_TEXTURE_2D, 0,
-			0, 0,
-			width, height,
-			GL_RGBA, GL_UNSIGNED_BYTE, data);
-		GL::CheckError();
-
-		Unbind();
-	}
-
-	void GenerateMipMaps()
-	{
-		// Create MipMap levels
-		Bind();
-
-		glHint(GL_GENERATE_MIPMAP_HINT, GL_NICEST);
-		GL::CheckError();
-
-		glGenerateMipmap(GL_TEXTURE_2D);
-		GL::CheckError();
-
-		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR_MIPMAP_LINEAR);
-		GL::CheckError();
-
-		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-		GL::CheckError();
-
-		Unbind();
-	}
+	void Bind();
+	void Unbind();
+	void WriteData(void* data);
+	void GenerateMipMaps();
 };
+
 typedef std::shared_ptr<Texture2D> Texture2DSPTR;
